+++ src/constants/path.ts
@@ -1,5 +1,7 @@
 export const PATH = {
   HOME: '/',
+  SIGN_UP: '/sign-up',
+  LOGIN: '/login',
   VENUES: '/venues',
   VENUES_DETAIL: '/venues/:venueId',
   VENUES_POST: '/venues/post',
