+++ src/pages/LoginPage/index.tsx
@@ -1,0 +1,3 @@
+export const LoginPage: React.FC = () => {
+  return <></>;
+};
