+++ src/pages/SignUpPage/index.tsx
@@ -1,0 +1,3 @@
+export const SignUpPage: React.FC = () => {
+  return <></>;
+};
